Rename application to lobstersgram
