+++ src/lobstersgram/__init__.py
@@ -1,0 +1,1 @@
+"""Lobstersgram: Lobsters → Telegraph → Telegram."""
