+++ src/lobstersgram/main.py
@@ -11,11 +11,11 @@ from pathlib import Path
 import feedparser
 from markdown_this import extract_main_content, fetch_url
 
-from lobstergram import config
-from lobstergram.content import Item, collect_new_items
-from lobstergram.state import load_state, load_subscribers, save_state, update_message_map
-from lobstergram.telegram import read_new_subscribers, resolve_recipient_chat_ids, send_to_recipients
-from lobstergram.telegraph import telegraph_create_page
+from lobstersgram import config
+from lobstersgram.content import Item, collect_new_items
+from lobstersgram.state import load_state, load_subscribers, save_state, update_message_map
+from lobstersgram.telegram import read_new_subscribers, resolve_recipient_chat_ids, send_to_recipients
+from lobstersgram.telegraph import telegraph_create_page
 
 
 class URLPublishError(RuntimeError):
